Rough Draft - exploring the data
Scatter Plots or age vs TSH, T3, TT4, and sex.
Looking for anything interesting to further explore.
Questions came up:
What time point in the patient history is each measurement from?  E.g. are all these measurements pre-treatment or post-treatment and if the latter at what point in the course of treatment (early or established?).  
Why are patients categorized as 'negative' in the Hypothyroid column on thyroxine?  Are these patients "cured" by the meds?  Or are they subclinical for hypothryoid and taking meds as a precaution?